fix #70
